refactor(frontend): promote shared table primitives to ui
