+++ apps/frontend/src/lib/components/admin/index.ts
@@ -1,6 +1,4 @@
 export { default as StatCard } from './StatCard.svelte';
-export { default as Panel } from './Panel.svelte';
-export { default as DataTable } from './DataTable.svelte';
 export { default as UserList } from './UserList.svelte';
 export { default as CopyId } from './CopyId.svelte';
 export { formatBytes, formatNumber } from './format';
