+++ apps/frontend/src/lib/components/rbac/PermissionMatrix.svelte
@@ -26,7 +26,7 @@ focusing a cell highlights its permission row and role column.
 -->
 <script lang="ts">
   import { onDestroy, type Snippet } from 'svelte';
-  import { Panel } from '$lib/components/admin';
+  import Panel from '$lib/ui/Panel.svelte';
   import { MatrixColumnHeading, MatrixTable } from '$lib/components/matrix';
   import { Hint, HelpTooltip } from '$lib/ui';
   import { ShortcutTextInput } from '$lib/ui/form';
