+++ apps/frontend/src/lib/ui/Panel.svelte.spec.ts
