+++ apps/frontend/src/lib/ui/index.ts
@@ -5,6 +5,7 @@ export { default as ChoiceRow } from './ChoiceRow.svelte';
 export { default as CollapsibleGroup } from './CollapsibleGroup.svelte';
 export { default as ConfirmDialog } from './ConfirmDialog.svelte';
 export { default as ContextMenu } from './ContextMenu.svelte';
+export { default as DataTable } from './DataTable.svelte';
 export { default as Dialog } from './Dialog.svelte';
 export { default as Divider } from './Divider.svelte';
 export { default as EmptyState } from './EmptyState.svelte';
@@ -23,6 +24,7 @@ export { default as MarkdownHtml } from './MarkdownHtml.svelte';
 export { default as MotdContent } from './MotdContent.svelte';
 export { default as NotificationBadge } from './NotificationBadge.svelte';
 export { default as PageTitle } from './PageTitle.svelte';
+export { default as Panel } from './Panel.svelte';
 export { default as PaneHeader } from './PaneHeader.svelte';
 export { default as PaneContent } from './PaneContent.svelte';
 export { default as Pill } from './Pill.svelte';
