+++ apps/frontend/src/routes/chat/[serverId]/manage/server/members/[userId]/MemberIdentitySettings.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { Panel } from '$lib/components/admin';
+  import Panel from '$lib/ui/Panel.svelte';
   import { m } from '$lib/i18n/messages';
   import { getLocale } from '$lib/i18n/runtime';
   import { useServerScope } from '$lib/state/server/scope.svelte';
