+++ apps/frontend/src/routes/chat/[serverId]/manage/server/members/[userId]/MemberRoleAssignments.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { resolve } from '$app/paths';
-  import { Panel } from '$lib/components/admin';
+  import Panel from '$lib/ui/Panel.svelte';
   import { m } from '$lib/i18n/messages';
   import { serverIdToSegment } from '$lib/navigation';
   import { Checkbox } from '$lib/ui/form';
