+++ apps/frontend/src/routes/chat/[serverId]/manage/server/permissions/new/role-create.page.svelte.spec.ts
@@ -33,8 +33,8 @@ vi.mock('$lib/state/server/scope.svelte', () => ({
   })
 }));
 vi.mock('$lib/api-client/roles', () => ({ createRoleAPI: vi.fn() }));
-vi.mock('$lib/components/admin', async () => ({
-  Panel: (await import('../[name]/RolePageSnippetMock.svelte')).default
+vi.mock('$lib/ui/Panel.svelte', async () => ({
+  default: (await import('../[name]/RolePageSnippetMock.svelte')).default
 }));
 vi.mock('$lib/ui', async () => ({
   PaneContent: (await import('../[name]/RolePageSnippetMock.svelte')).default
