+++ apps/frontend/src/routes/chat/[serverId]/manage/server/system/AssetCleanupPanel.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import type { AdminAssetCleanupStatus } from '$lib/api-client/adminDiagnostics';
-  import { Panel } from '$lib/components/admin';
+  import Panel from '$lib/ui/Panel.svelte';
   import { Pill } from '$lib/ui';
   import { m } from '$lib/i18n/messages';
