+++ apps/frontend/src/routes/chat/[serverId]/settings/ProfileDetailsSettings.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { useServerScope } from '$lib/state/server/scope.svelte';
   import type { AccountAPI } from '$lib/api-client/account';
-  import { Panel } from '$lib/components/admin';
+  import Panel from '$lib/ui/Panel.svelte';
   import { m } from '$lib/i18n/messages';
   import { Dialog, Hint } from '$lib/ui';
   import { Button, Form, TextArea, TextInput } from '$lib/ui/form';
