+++ apps/frontend/src/routes/chat/[serverId]/settings/account/PasswordSettings.svelte
@@ -2,7 +2,7 @@
   import { Code, ConnectError } from '@connectrpc/connect';
   import type { AccountAPI } from '$lib/api-client/account';
   import type { CurrentUserState } from '$lib/auth/currentUser.svelte';
-  import { Panel } from '$lib/components/admin';
+  import Panel from '$lib/ui/Panel.svelte';
   import { m } from '$lib/i18n/messages';
   import { Button, FormError, TextInput, validate, z } from '$lib/ui/form';
   import { toast } from '$lib/ui/toast/toastState.svelte';
